total management apis updated

+++ app/Services/TotalManagementExpensesServices.php
@@ -129,8 +129,13 @@ class TotalManagementExpensesServices
     {
         return $this->totalManagementExpenses::with(['totalManagementExpensesAttachments' => function ($query) {
             $query->orderBy('created_at', 'desc');
-        }])->leftJoin('workers', 'workers.id', 'total_management_expenses.worker_id')
-        ->whereIn('workers.company_id', $request['company_id'])->findOrFail($request['id']);
+        }])
+        ->join('workers', function ($join) use ($request) {
+            $join->on('workers.id', '=', 'total_management_expenses.worker_id')
+                ->whereIn('workers.company_id', $request['company_id']);
+        })
+        ->select('total_management_expenses.*')
+        ->find($request['id']);
     }
     /**
      * @param $request
@@ -197,9 +202,16 @@ class TotalManagementExpensesServices
         $params['modified_by'] = $user['id'];
         $params['company_id'] = $this->authServices->getCompanyIds($user);
         $expense = $this->totalManagementExpenses
-        ->leftJoin('workers', 'workers.id', 'total_management_expenses.worker_id')
-        ->whereIn('workers.company_id', $params['company_id'])
-        ->findOrFail($request['id']);
+        ->join('workers', function ($join) use ($params) {
+            $join->on('workers.id', '=', 'total_management_expenses.worker_id')
+                ->whereIn('workers.company_id', $params['company_id']);
+        })
+        ->find($request['id']);
+        if(is_null($expense)){
+            return [
+                'noRecords' => true
+            ];
+        }
         $expense->worker_id = $params['worker_id'] ?? $expense->worker_id;
         $expense->application_id = $params['application_id'] ?? $expense->application_id;
         $expense->project_id = $params['project_id'] ?? $expense->project_id;
@@ -237,8 +249,17 @@ class TotalManagementExpensesServices
      */
     public function delete($request) : bool
     {
-        $expense = $this->totalManagementExpenses->leftJoin('workers', 'workers.id', 'total_management_expenses.worker_id')
-        ->whereIn('workers.company_id', $request['company_id'])->findOrFail($request['id']);
+        $expense = $this->totalManagementExpenses
+        ->join('workers', function ($join) use ($request) {
+            $join->on('workers.id', '=', 'total_management_expenses.worker_id')
+                ->whereIn('workers.company_id', $request['company_id']);
+        })
+        ->find($request['id']);
+
+        if(is_null($expense)){
+            return false;
+        }
+
         $expense->delete();
         return true;
     }
@@ -249,7 +270,15 @@ class TotalManagementExpensesServices
      */    
     public function deleteAttachment($request): bool
     {   
-        $data = $this->totalManagementExpensesAttachments::find($request['id']); 
+        $data = $this->totalManagementExpensesAttachments::join('total_management_expenses', 'total_management_expenses.id', 'total_management_expenses_attachments.file_id')
+        ->join('workers', function ($join) use ($request) {
+            $join->on('workers.id', '=', 'total_management_expenses.worker_id')
+                ->whereIn('workers.company_id', $request['company_id']);
+        })
+        ->find($request['id']); 
+        if(is_null($data)){
+            return false;
+        }
         $data->delete();
         return true;
     }
